requirements.txt
fastapi
uvicorn
uvicorn[standard]
python-multipart
python-docx
langchain[faiss]
openai
faiss-cpu
langchain-community
#dotenv
tiktoken